Create Energy Consumptions Dashboard PowerBI
This in an in-depth Power BI visualisation regarding the consumption of different energy sources from various areas within the United States. This visualisation was created through the use of PowerQuery, DAX and other functions.This in an in-depth Power BI visualisation regarding the consumption of different energy sources from various areas within the United States. This visualisation was created through the use of PowerQuery, DAX and other functions.

Skills: Microsoft Power Query · Microsoft Power BI · Data Pipelines

Please take a look at the link below: 

https://app.powerbi.com/view?r=eyJrIjoiZTc3YjQ5OGMtYTMyOC00OGVhLWJjZTUtYWY5NzhjZTNhODdlIiwidCI6ImFlZmU3YTAzLWRiYmQtNDI4YS1iNjdhLWYwYjZhZmVhZTA2YyJ9